gl_draw : Accept float coordinates and sizes for better precision
Also staticificatons, constifications

+++ Quake/draw.h
@@ -30,16 +30,17 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 extern qpic_t *draw_disc; // also used on sbar
 
+// Use float coords and sizes in Draw_XXX to allow sub-pixel precision
 void	Draw_Init (void);
-void	Draw_Character (cb_context_t *cbx, int x, int y, int num);
-void	Draw_Pic (cb_context_t *cbx, int x, int y, qpic_t *pic, float alpha, qboolean alpha_blend);
+void	Draw_Character (cb_context_t *cbx, float x, float y, int num);
+void	Draw_Pic (cb_context_t *cbx, float x, float y, qpic_t *pic, float alpha, qboolean alpha_blend);
 void	Draw_SubPic (cb_context_t *cbx, float x, float y, float w, float h, qpic_t *pic, float s1, float t1, float s2, float t2, float *rgb, float alpha);
-void	Draw_TransPicTranslate (cb_context_t *cbx, int x, int y, qpic_t *pic, int top, int bottom); // johnfitz -- more parameters
-void	Draw_ConsoleBackground (cb_context_t *cbx);													// johnfitz -- removed parameter int lines
-void	Draw_TileClear (cb_context_t *cbx, int x, int y, int w, int h);
-void	Draw_Fill (cb_context_t *cbx, int x, int y, int w, int h, int c, float alpha); // johnfitz -- added alpha
+void	Draw_TransPicTranslate (cb_context_t *cbx, float x, float y, qpic_t *pic, int top, int bottom); // johnfitz -- more parameters
+void	Draw_ConsoleBackground (cb_context_t *cbx);														// johnfitz -- removed parameter int lines
+void	Draw_TileClear (cb_context_t *cbx, float x, float y, float w, float h);
+void	Draw_Fill (cb_context_t *cbx, float x, float y, float w, float h, int c, float alpha); // johnfitz -- added alpha
 void	Draw_FadeScreen (cb_context_t *cbx);
-void	Draw_String (cb_context_t *cbx, int x, int y, const char *str);
+void	Draw_String (cb_context_t *cbx, float x, float y, const char *str);
 void	Draw_String_3D (cb_context_t *cbx, vec3_t coords, float size, const char *str);
 qpic_t *Draw_PicFromWad2 (const char *name, unsigned int texflags);
 qpic_t *Draw_PicFromWad (const char *name);
